requirements.txt
streamlit
mysql-connector-python
python-dotenv
boto3
spacy
pdfplumber
python-docx
